requirements.txt
bezier==2021.2.12
click==8.1.3
colorama==0.4.5
cycler==0.11.0
Flask==2.2.2
fonttools==4.37.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.20.3
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
scipy==1.7.3
six==1.16.0
typing-extensions==4.3.0
Werkzeug==2.2.2
zipp==3.8.1
